Bugfixed struct iovec fallback.

+++ src/modules/files/file.h
@@ -52,7 +52,7 @@ struct my_file
 #ifndef HAVE_STRUCT_IOVEC
 struct iovec {
   void *iov_base;
-  int iov_len;
+  size_t iov_len;
 };
 #endif /* !HAVE_STRUCT_IOVEC */
